requirements.txt
flask
flask-cors
requests
ddgs
Pillow
python-dotenv
